weather_bot/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via httpcore
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==2.1.1
    # via requests
h11==0.14.0
    # via httpcore
httpcore==0.16.3
    # via httpx
httpx==0.23.1
    # via python-telegram-bot
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
psycopg2-binary==2.9.5
    # via -r requirements.in
python-telegram-bot==20.0b0
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
requests==2.28.1
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
urllib3==1.26.13
    # via requests
